fix: wrap publications in suspense boundary on admin statistics page
